# backend/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Inscription API
    version: 0.0.1
paths:
    /api/follow:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetFollower
            parameters:
                - name: address
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.GetFollowerResponse'
    /api/follow_tweet:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetFollowTweet
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: offset
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.TweetResponse'
    /api/group:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetGroup
            parameters:
                - name: address
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.SwiftResponses'
    /api/group_message:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetGroupMessage
            parameters:
                - name: title
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: offset
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.GetMessageResponse'
    /api/message:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetMessage
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: to
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: offset
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.GetMessageResponse'
    /api/message/{hash}:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetMessageByHash
            parameters:
                - name: hash
                  in: path
                  required: true
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
    /api/message_window:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetMessageWindow
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.MessageWindow'
    /api/profile:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetProfile
            parameters:
                - name: address
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
    /api/tweet:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetTweet
            parameters:
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: offset
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.TweetResponse'
    /api/tweet/{address}:
        get:
            tags:
                - Inscription
            operationId: Inscription_GetTweetByAddress
            parameters:
                - name: address
                  in: path
                  required: true
                  schema:
                    type: string
                - name: owner
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: offset
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inscription.v1.TweetResponse'
components:
    schemas:
        api.inscription.v1.GetFollowerResponse:
            type: object
            properties:
                follower:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
        api.inscription.v1.GetMessageResponse:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
        api.inscription.v1.MessageWindow:
            type: object
            properties:
                profiles:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
        api.inscription.v1.SwiftResponse:
            type: object
            properties:
                type:
                    type: string
                title:
                    type: string
                text:
                    type: string
                image:
                    type: array
                    items:
                        type: string
                receiver:
                    type: array
                    items:
                        type: string
                at:
                    type: array
                    items:
                        type: string
                with:
                    type: string
                height:
                    type: integer
                    format: int64
                trxHash:
                    type: string
                trxTime:
                    type: string
                sender:
                    type: string
        api.inscription.v1.SwiftResponses:
            type: object
            properties:
                groups:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
        api.inscription.v1.TweetResponse:
            type: object
            properties:
                tweets:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inscription.v1.Tweets'
        api.inscription.v1.Tweets:
            type: object
            properties:
                tweet:
                    $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
                with:
                    $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inscription.v1.SwiftResponse'
                likeNum:
                    type: integer
                    format: int64
                likeBool:
                    type: boolean
tags:
    - name: Inscription